+++ packages/frontend/workspace/src/atom.ts
@@ -74,15 +74,12 @@ const rootWorkspacesMetadataPrimitiveAtom = atom<Promise<
 
 type Getter = <Value>(atom: Atom<Value>) => Value;
 
-type FetchMetadata = (
-  get: Getter,
-  options: { signal: AbortSignal }
-) => Promise<RootWorkspaceMetadata[]>;
+type FetchMetadata = (get: Getter) => Promise<RootWorkspaceMetadata[]>;
 
 /**
  * @internal
  */
-const fetchMetadata: FetchMetadata = async (get, { signal }) => {
+const fetchMetadata: FetchMetadata = async get => {
   performanceJotaiLogger.info('fetch metadata start');
 
   const WorkspaceAdapters = get(workspaceAdaptersAtom);
@@ -111,23 +108,6 @@ const fetchMetadata: FetchMetadata = async (get, { signal }) => {
       }
       return [];
     };
-
-    const maybeMetadata = loadFromLocalStorage();
-
-    // migration step, only data in `METADATA_STORAGE_KEY` will be migrated
-    if (
-      maybeMetadata.some(meta => !('version' in meta)) &&
-      !window.$migrationDone
-    ) {
-      await new Promise<void>((resolve, reject) => {
-        signal.addEventListener('abort', () => reject(), { once: true });
-        window.addEventListener('migration-done', () => resolve(), {
-          once: true,
-        });
-      });
-      performanceJotaiLogger.info('migration done');
-    }
-
     metadata.push(...loadFromLocalStorage());
   }
   // step 2: fetch from adapters
@@ -211,14 +191,14 @@ const fetchMetadata: FetchMetadata = async (get, { signal }) => {
 
 const rootWorkspacesMetadataPromiseAtom = atom<
   Promise<RootWorkspaceMetadata[]>
->(async (get, { signal }) => {
+>(async get => {
   const primitiveMetadata = get(rootWorkspacesMetadataPrimitiveAtom);
   assertEquals(
     primitiveMetadata,
     null,
     'rootWorkspacesMetadataPrimitiveAtom should be null'
   );
-  return fetchMetadata(get, { signal });
+  return fetchMetadata(get);
 });
 
 type SetStateAction<Value> = Value | ((prev: Value) => Value);
@@ -276,11 +256,7 @@ export const rootWorkspacesMetadataAtom = atom<
 );
 
 export const refreshRootMetadataAtom = atom(null, (get, set) => {
-  const abortController = new AbortController();
-  set(
-    rootWorkspacesMetadataPrimitiveAtom,
-    fetchMetadata(get, { signal: abortController.signal })
-  );
+  set(rootWorkspacesMetadataPrimitiveAtom, fetchMetadata(get));
 });
 
 // blocksuite atoms,
