+++ tests/affine-migration/README.md
@@ -7,3 +7,6 @@ BUILD_TYPE=canary yarn run build
 cd tests/affine-migration
 yarn run e2e
 ```
+
+> Tips:
+> Run `yarn dev` to start dev server in 8080 could make debugging more quickly.
