+++ tests/kit/utils/workspace.ts
@@ -30,10 +30,13 @@ export async function createLocalWorkspace(
   await page.getByPlaceholder('Set a Workspace name').fill(params.name);
 
   // click create button
-  await page.getByRole('button', { name: 'Create' }).click({
+  await page.getByTestId('create-workspace-create-button').click({
     delay: 500,
   });
 
+  await expect(
+    page.getByTestId('create-workspace-create-button')
+  ).not.toBeAttached();
   await waitForEditorLoad(page);
 
   await expect(page.getByTestId('workspace-name')).toHaveText(params.name);
